requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
jinja2==3.1.5
aiofiles==24.1.0
redis==5.2.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
jsonschema==4.23.0
cryptography==44.0.0
stripe==11.4.1
slowapi==0.1.9
anthropic==0.42.0
